% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_add.R
\name{add}
\alias{add}
\title{Calculation of Sum}
\usage{
add(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
sum of two numbers
}
\description{
Calculation of Sum
}
\examples{
add(1, 2)
}
